Add styling to the navbar - Refactor Navigation component
